+++ build_utils/mdl_style.rb
@@ -1,0 +1,5 @@
+all
+rule 'MD007', :indent => 4
+rule 'MD013', :line_length => 500
+rule 'MD029', :style => :ordered
+rule "MD030", :ul_multi => 3, :ol_multi => 2, :ol_single => 2, :ul_single => 3