+++ package/PartSeg/common_gui/searchable_list_widget.py
@@ -22,10 +22,10 @@ class SearchableListWidget(QWidget):
         return super().__getattr__(item)
 
     def update_visible(self, text):
-        items = self.list_widget.findItems(text, Qt.MatchContains)
+        items_text = [x.text() for x in self.list_widget.findItems(text, Qt.MatchContains)]
         for index in range(self.list_widget.count()):
             item = self.item(index)
-            item.setHidden(item not in items)
+            item.setHidden(item.text() not in items_text)
 
     def addItems(self, *args):
         self.list_widget.addItems(*args)